+++ app/(home)/petAccount/layout.tsx
@@ -1,9 +1,9 @@
-import Heading from '@/components/Account/Heading';
+import Heading from "@/components/Account/Heading";
 
 const AccountLayout = ({ children }: { children: React.ReactNode }) => {
   return (
     <div>
-      <Heading title='회원가입' />
+      <Heading title="회원가입" />
       {children}
     </div>
   );
