+++ app/(home)/petAccount/page.tsx
@@ -1,11 +1,9 @@
 "use client";
 
-import { PetForm } from "@/components/PetAccount/Heading/petForm";
-
 const PetAccount = () => {
   return (
     <div>
-      <PetForm />
+      <PetAccount />
     </div>
   );
 };
