+++ app/components/PetAccount/TypeButton.tsx
